+++ runtime/outcpp/System.Private.CoreLib/System/Reflection/Emit/PunkSafeHandle.cpp
@@ -1,5 +1,6 @@
 #include "PunkSafeHandle-dep.h"
 
+#include <System.Private.CoreLib/System/Int32-dep.h>
 #include <System.Private.CoreLib/System/Reflection/Emit/PunkSafeHandle-dep.h>
 #include <System.Private.CoreLib/System/Runtime/InteropServices/Marshal-dep.h>
 #include <System.Private.CoreLib/System/Type-dep.h>
@@ -8,7 +9,7 @@ namespace System::Private::CoreLib::System::Reflection::Emit::PunkSafeHandleName
 using namespace System::Runtime::InteropServices;
 
 Boolean PunkSafeHandle___::get_IsInvalid() {
-  return handle == (IntPtr)0;
+  return handle == (IntPtr)(Int32)0;
 }
 
 Boolean PunkSafeHandle___::ReleaseHandle() {
@@ -18,7 +19,7 @@ Boolean PunkSafeHandle___::ReleaseHandle() {
 
 void PunkSafeHandle___::cctor() {
   m_Release = &(DRelease)Marshal::GetDelegateForFunctionPointer(nGetDReleaseTarget(), typeof<DRelease>());
-  m_Release((IntPtr)0);
+  m_Release((IntPtr)(Int32)0);
 }
 
 } // namespace System::Private::CoreLib::System::Reflection::Emit::PunkSafeHandleNamespace
