+++ runtime/outcpp/System.Private.CoreLib/System/Text/UTF7Encoding.cpp
@@ -398,7 +398,7 @@ Int32 UTF7Encoding___::GetBytes(Char* chars, Int32 charCount, Byte* bytes, Int32
       continue;
     }
     if (num2 < 0 && nextChar == 43) {
-      if (!encodingByteBuffer->AddByte((?)43, (?)45)) {
+      if (!encodingByteBuffer->AddByte((?)(Int32)43, (?)(Int32)45)) {
         break;
       }
       continue;
